+++ webdevictt/history.html
@@ -44,7 +44,7 @@
                 </div>
                 <div class="menu-card">
                     <img src="images/Group 21.svg" alt="">
-                    <p>Support</p>
+                    <a href="mailto:demiladeadelekun2020@gmail.com" style="color: grey;"> <p>Support</p></a>
                 </div>
             </div>
             <div class="logout-nav">
