+++ webdevictt/regulations.html
@@ -43,7 +43,7 @@
                 </div>
                 <div class="menu-card">
                     <img src="images/Group 21.svg" alt="">
-                    <p>Support</p>
+                    <a href="mailto:demiladeadelekun2020@gmail.com" style="color: grey;"> <p>Support</p></a>
                 </div>
             </div>
             <div class="logout-nav">
@@ -72,7 +72,7 @@
             </div>
         </section>
     </main>
-    <script type="module" src="rules.js"></script>
+    <script type="module" src="rule.js"></script>
     <script src="dashboard.js" type="module"></script>
 </body>
 </html>
